Changes:  - removed redundant include directives
git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@6332 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ libraries/trace/source/trace.cpp
@@ -282,7 +282,7 @@ ATerm readATerm(istream &is)
 	}
         is.clear();
 
-	ATerm t = ATreadFromBinaryString((unsigned char *) buf,len);
+	ATerm t = ATreadFromBinaryString((unsigned char *) buf, static_cast< int >(len));
         if ( t == NULL )
         {
           throw runtime_error("failed to read ATerm from stream");
